9C-969: Moves CacheWrapper class to commons
