+++ project/Dependencies.scala
@@ -58,10 +58,15 @@ object Dependencies {
 
   val commonsDeps = Seq(libraryDependencies ++= Seq(
     cats,
-    scalaz("-concurrent"),
+    circe("-core"),
+    circe("-generic"),
+    circe("-parser"),
+    embeddedRedis,
     enumeratum(""),
     jodaConvert,
     jodaTime,
+    redisClient,
+    scalaz("-concurrent"),
     specs2Core,
     specs2("-scalacheck"),
     scalacheckShapeless,
